@@ -1,59 +1,59 @@
 {
     "name": "los/api-server",
     "description": "PHP api server middleware",
+    "license": "MIT",
+    "type": "library",
+    "authors": [
+        {
+            "name": "Leandro Silva",
+            "homepage": "https://sillotec.com/"
+        }
+    ],
+    "homepage": "https://github.com/lansoweb/api-server",
     "require": {
-        "php": "^7.4 | ^8.0",
+        "php": "^8.0",
         "ext-json": "*",
-        "psr/container": "^1.0 | ^2.0",
-        "psr/http-message": "^1.0",
-        "laminas/laminas-inputfilter": "^2.7",
-        "laminas/laminas-paginator": "^2.7",
+        "laminas/laminas-diactoros": "^3.2",
         "laminas/laminas-eventmanager": "^3.0",
-        "psr/http-server-middleware": "^1.0",
+        "laminas/laminas-inputfilter": "^2.27",
+        "laminas/laminas-paginator": "^2.7",
+        "los/uql": "^1.1",
         "mezzio/mezzio-hal": "^2.0",
         "mezzio/mezzio-helpers": "^5.6",
         "mezzio/mezzio-problem-details": "^1.0",
-        "laminas/laminas-diactoros": "^1.7 || ^2.2",
-        "los/uql": "^1.1",
-        "doctrine/coding-standard": "^9.0"
+        "psr/container": "^1.0 || ^2.0",
+        "psr/http-message": "^1.0 || ^2.0",
+        "psr/http-server-middleware": "^1.0"
     },
     "require-dev": {
-        "squizlabs/php_codesniffer": "^3.6",
-        "phpstan/phpstan": "^0.12",
+        "doctrine/coding-standard": "^12.0",
         "laminas/laminas-db": "^2.10",
-        "laminas/laminas-hydrator": "^4.0"
-    },
-    "license": "MIT",
-    "autoload-dev": {
-        "psr-4": {
-            "LosMiddleware\\ApiServerTest\\": "test/"
-        }
+        "laminas/laminas-hydrator": "^4.0",
+        "phpstan/phpstan": "^1.10",
+        "squizlabs/php_codesniffer": "^3.6"
     },
     "autoload": {
         "psr-4": {
-            "LosMiddleware\\ApiServer\\": "src/"
+            "Los\\ApiServer\\": "src/"
         }
     },
-    "type": "library",
+    "autoload-dev": {
+        "psr-4": {
+            "Los\\ApiServerTest\\": "test/"
+        }
+    },
+    "config": {
+        "allow-plugins": {
+            "dealerdirect/phpcodesniffer-composer-installer": true
+        }
+    },
     "scripts": {
-        "cs-check": "phpcs",
-        "cs-fix": "phpcbf",
         "check": [
             "@cs-check",
             "@phpstan"
         ],
-        "phpstan": "phpstan analyse -l 4 -c phpstan.neon src"
-    },
-    "homepage": "https://github.com/lansoweb/api-server",
-    "authors": [
-        {
-            "name": "Leandro Silva",
-            "homepage": "http://leandrosilva.info/"
-        }
-    ],
-    "extra": {
-        "zf": {
-            "config-provider": "LosMiddleware\\ApiServer\\ConfigProvider"
-        }
+        "cs-check": "phpcs",
+        "cs-fix": "phpcbf",
+        "phpstan": "phpstan analyse"
     }
 }
