@@ -1,52 +1,62 @@
 <?php
-declare(strict_types = 1);
 
-namespace LosMiddleware\ApiServer\Handler;
+declare(strict_types=1);
 
-use LosMiddleware\ApiServer\Entity\Collection;
-use LosMiddleware\ApiServer\Entity\Entity;
-use LosMiddleware\ApiServer\Entity\EntityInterface;
-use LosMiddleware\ApiServer\Exception\NotFoundException;
-use LosMiddleware\ApiServer\Mapper\MapperInterface;
+namespace Los\ApiServer\Handler;
+
+use Los\ApiServer\Entity\Collection;
+use Los\ApiServer\Entity\Entity;
+use Los\ApiServer\Entity\EntityInterface;
+use Los\ApiServer\Exception\NotFoundException;
+use Los\ApiServer\Mapper\MapperInterface;
 use Mezzio\Hal\HalResponseFactory;
 use Mezzio\Hal\ResourceGenerator;
 use Mezzio\Helper\UrlHelper;
 use Mezzio\ProblemDetails\ProblemDetailsResponseFactory;
 
+use function array_key_exists;
+use function array_merge;
+use function end;
+use function explode;
+use function json_decode;
+use function str_replace;
+use function strtolower;
+
 abstract class MapperRestHandler extends AbstractRestHandler
 {
-    const SORT_BY = self::IDENTIFIER_NAME;
+    public const SORT_BY = self::IDENTIFIER_NAME;
 
-    protected MapperInterface $mapper;
     protected int $limitItemsPerPage = 25;
 
     public function __construct(
-        MapperInterface $mapper,
+        protected MapperInterface $mapper,
         Entity $entityPrototype,
         UrlHelper $urlHelper,
         ProblemDetailsResponseFactory $problemDetailsResponseFactory,
         ResourceGenerator $resourceGenerator,
-        HalResponseFactory $responseFactory
+        HalResponseFactory $responseFactory,
     ) {
         parent::__construct($urlHelper, $problemDetailsResponseFactory, $resourceGenerator, $responseFactory);
 
-        $this->mapper = $mapper;
         $this->entityPrototype = $entityPrototype;
     }
 
     /**
      * {@inheritDoc}
-     * @see \LosMiddleware\ApiServer\Handler\AbstractRestHandler::getResourceName()
+     *
+     * @see \Los\ApiServer\Handler\AbstractRestHandler::getResourceName()
      */
     public function getResourceName(): string
     {
-        $tokens = explode('\\', get_class($this));
+        $tokens = explode('\\', static::class);
+
         return strtolower(str_replace('Handler', '', end($tokens)));
     }
 
     /**
      * {@inheritDoc}
-     * @see \LosMiddleware\ApiServer\Handler\AbstractRestHandler::create()
+     *
+     * @see \Los\ApiServer\Handler\AbstractRestHandler::create()
      */
     public function create(array $data): EntityInterface
     {
@@ -61,37 +71,40 @@ abstract class MapperRestHandler extends AbstractRestHandler
 
     /**
      * {@inheritDoc}
-     * @see \LosMiddleware\ApiServer\Handler\AbstractRestHandler::fetch()
+     *
+     * @see \Los\ApiServer\Handler\AbstractRestHandler::fetch()
      */
     public function fetch($id, array $where = []): EntityInterface
     {
-        $where = array_merge([static::IDENTIFIER_NAME => $id], $where);
+        $where = array_merge([self::IDENTIFIER_NAME => $id], $where);
 
         $entity = $this->mapper->findOneBy($where);
         if ($entity === null) {
             throw NotFoundException::create();
         }
 
-        $query = $this->request->getQueryParams();
+        $query  = $this->request->getQueryParams();
         $fields = $query['fields'] ?? '';
         if (! empty($fields)) {
             $entity->setFields(explode(',', $fields));
         }
+
         return $entity;
     }
 
     /**
      * {@inheritDoc}
-     * @see \LosMiddleware\ApiServer\Handler\AbstractRestHandler::fetchAll()
+     *
+     * @see \Los\ApiServer\Handler\AbstractRestHandler::fetchAll()
      */
     public function fetchAll(array $where = [], array $options = []): Collection
     {
         $queryParams = $this->request->getQueryParams();
-        $query = array_key_exists('q', $queryParams) ? json_decode($queryParams['q'], true) : [];
-        $hint = array_key_exists('h', $queryParams) ? json_decode($queryParams['h'], true) : [];
+        $query       = array_key_exists('q', $queryParams) ? json_decode($queryParams['q'], true) : [];
+        $hint        = array_key_exists('h', $queryParams) ? json_decode($queryParams['h'], true) : [];
 
         $where = array_merge($where, $query);
-        $hint = array_merge($options, $hint);
+        $hint  = array_merge($options, $hint);
 
         $collection = $this->mapper->findBy($where, $hint);
 
@@ -110,11 +123,12 @@ abstract class MapperRestHandler extends AbstractRestHandler
 
     /**
      * {@inheritDoc}
-     * @see \LosMiddleware\ApiServer\Handler\AbstractRestHandler::delete()
+     *
+     * @see \Los\ApiServer\Handler\AbstractRestHandler::delete()
      */
-    public function delete($id, array $where = [])
+    public function delete($id, array $where = []): void
     {
-        $where = array_merge([static::IDENTIFIER_NAME => $id], $where);
+        $where = array_merge([self::IDENTIFIER_NAME => $id], $where);
 
         $entity = $this->mapper->findOneBy($where);
         if ($entity === null) {
@@ -126,11 +140,12 @@ abstract class MapperRestHandler extends AbstractRestHandler
 
     /**
      * {@inheritDoc}
-     * @see \LosMiddleware\ApiServer\Handler\AbstractRestHandler::patch()
+     *
+     * @see \Los\ApiServer\Handler\AbstractRestHandler::patch()
      */
     public function patch($id, array $data, array $where = []): EntityInterface
     {
-        $where = array_merge([static::IDENTIFIER_NAME => $id], $where);
+        $where = array_merge([self::IDENTIFIER_NAME => $id], $where);
 
         $entity = $this->mapper->findOneBy($where);
         if ($entity === null) {
@@ -147,11 +162,12 @@ abstract class MapperRestHandler extends AbstractRestHandler
 
     /**
      * {@inheritDoc}
-     * @see \LosMiddleware\ApiServer\Handler\AbstractRestHandler::update()
+     *
+     * @see \Los\ApiServer\Handler\AbstractRestHandler::update()
      */
     public function update($id, array $data, array $where = []): EntityInterface
     {
-        $where = array_merge([static::IDENTIFIER_NAME => $id], $where);
+        $where = array_merge([self::IDENTIFIER_NAME => $id], $where);
 
         $entity = $this->mapper->findOneBy($where);
         if ($entity === null) {
