@@ -1,28 +1,25 @@
 <?php
-declare(strict_types = 1);
 
-namespace LosMiddleware\ApiServer\Paginator;
+declare(strict_types=1);
 
-use LosMiddleware\ApiServer\Mapper\MapperInterface;
+namespace Los\ApiServer\Paginator;
+
 use Laminas\Paginator\Adapter\AdapterInterface;
+use Los\ApiServer\Mapper\MapperInterface;
 
 class MapperAdapter implements AdapterInterface
 {
-    private MapperInterface $mapper;
-    private array $where;
-    private $order;
-    private $group;
-
-    public function __construct(MapperInterface $mapper, array $where = [], $order = null, $group = null)
-    {
-        $this->mapper = $mapper;
-        $this->where = $where;
-        $this->order = $order;
-        $this->group = $group;
+    public function __construct(
+        private MapperInterface $mapper,
+        private array $where = [],
+        private $order = null,
+        private $group = null,
+    ) {
     }
 
     /**
      * {@inheritDoc}
+     *
      * @see \Laminas\Paginator\Adapter\AdapterInterface::getItems()
      */
     public function getItems($offset, $itemCountPerPage)
@@ -37,9 +34,10 @@ class MapperAdapter implements AdapterInterface
 
     /**
      * {@inheritDoc}
+     *
      * @see Countable::count()
      */
-    public function count()
+    public function count(): int
     {
         return $this->mapper->count($this->where);
     }
