@@ -1,7 +1,8 @@
 <?php
-declare(strict_types = 1);
 
-namespace LosMiddleware\ApiServer\Entity;
+declare(strict_types=1);
+
+namespace Los\ApiServer\Entity;
 
 use Laminas\Paginator\Paginator;
 
