@@ -1,7 +1,8 @@
 <?php
-declare(strict_types = 1);
 
-namespace LosMiddleware\ApiServer\Exception;
+declare(strict_types=1);
+
+namespace Los\ApiServer\Exception;
 
 use Mezzio\ProblemDetails\Exception\CommonProblemDetailsExceptionTrait;
 use Mezzio\ProblemDetails\Exception\ProblemDetailsExceptionInterface;
@@ -10,13 +11,14 @@ class AuthorizationException extends RuntimeException implements ProblemDetailsE
 {
     use CommonProblemDetailsExceptionTrait;
 
-    public static function create(string $message = 'Not Authorized') : self
+    public static function create(string $message = 'Not Authorized'): self
     {
-        $exception = new self($message, 401);
+        $exception         = new self($message, 401);
         $exception->status = 401;
         $exception->detail = $message;
-        $exception->type = '';
-        $exception->title = '';
+        $exception->type   = '';
+        $exception->title  = '';
+
         return $exception;
     }
 }
