@@ -1,7 +1,8 @@
 <?php
-declare(strict_types = 1);
 
-namespace LosMiddleware\ApiServer\Exception;
+declare(strict_types=1);
+
+namespace Los\ApiServer\Exception;
 
 class RuntimeException extends \RuntimeException
 {
