@@ -1,7 +1,8 @@
 <?php
-declare(strict_types = 1);
 
-namespace LosMiddleware\ApiServer\Exception;
+declare(strict_types=1);
+
+namespace Los\ApiServer\Exception;
 
 use Mezzio\ProblemDetails\Exception\CommonProblemDetailsExceptionTrait;
 use Mezzio\ProblemDetails\Exception\ProblemDetailsExceptionInterface;
@@ -10,14 +11,15 @@ class ValidationException extends RuntimeException implements ProblemDetailsExce
 {
     use CommonProblemDetailsExceptionTrait;
 
-    public static function fromMessages(array $messages) : self
+    public static function fromMessages(array $messages): self
     {
-        $exception = new self('Unprocessable Entity', 422);
-        $exception->status = 422;
-        $exception->detail = 'Unprocessable Entity';
-        $exception->type = '';
-        $exception->title = '';
+        $exception             = new self('Unprocessable Entity', 422);
+        $exception->status     = 422;
+        $exception->detail     = 'Unprocessable Entity';
+        $exception->type       = '';
+        $exception->title      = '';
         $exception->additional = ['messages' => $messages];
+
         return $exception;
     }
 }
